+++ yash-prompt/CHANGELOG.md
@@ -13,12 +13,13 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
-## [0.15.1] - Unreleased
+## [0.16.0] - Unreleased
 
 ### Changed
 
 - Public dependency versions:
-    - yash-env 0.16.0 → 0.16.1
+    - yash-env 0.16.0 → 0.17.0
+    - yash-syntax 0.24.0 → 0.25.0
 
 ## [0.15.0] - 2026-07-31
 
@@ -206,7 +207,7 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 - Initial implementation of the `yash-prompt` crate
 
-[0.15.1]: https://github.com/magicant/yash-rs/releases/tag/yash-prompt-0.15.1
+[0.16.0]: https://github.com/magicant/yash-rs/releases/tag/yash-prompt-0.16.0
 [0.15.0]: https://github.com/magicant/yash-rs/releases/tag/yash-prompt-0.15.0
 [0.14.0]: https://github.com/magicant/yash-rs/releases/tag/yash-prompt-0.14.0
 [0.13.0]: https://github.com/magicant/yash-rs/releases/tag/yash-prompt-0.13.0