+++ yash-syntax/CHANGELOG.md
@@ -13,12 +13,12 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
-## [0.24.1] - Unreleased
+## [0.25.0] - Unreleased
 
 ### Changed
 
 - Public dependency versions:
-    - yash-env 0.16.0 → 0.16.1
+    - yash-env 0.16.0 → 0.17.0
 
 ## [0.24.0] - 2026-07-31
 
@@ -817,7 +817,7 @@ command.
 - Functionalities to parse POSIX shell scripts
 - Alias substitution support
 
-[0.24.1]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.24.1
+[0.25.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.25.0
 [0.24.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.24.0
 [0.23.1]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.23.1
 [0.23.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.23.0